(Refactor) settings.php
- refactored monolithic version of settings.php into different files
