+++ app/controllers/SettingsController.php
@@ -5,20 +5,32 @@ require_once __DIR__ . '/../../core/functions.php';
 
 class SettingsController {
 
-    private function checkAuth() {
+    private $db;
+    private $auth;
+
+    private function checkAuthAndInit() {
         if (!isset($_SESSION['user'])) {
-            header('Location: /iCensus-ent/public/login');
+            // For AJAX requests, return JSON. For page loads, redirect.
+            if (!empty($_SERVER['HTTP_X_REQUESTED_WITH']) && strtolower($_SERVER['HTTP_X_REQUESTED_WITH']) == 'xmlhttprequest') {
+                header('Content-Type: application/json');
+                http_response_code(401);
+                echo json_encode(['status' => 'error', 'message' => 'Unauthorized']);
+            } else {
+                header('Location: /iCensus-ent/public/login');
+            }
             exit;
         }
+        
+        // Initialize DB and Auth for other methods to use
+        $config = require __DIR__ . '/../../config/database.php';
+        $this->db = new Database($config);
+        $GLOBALS['db'] = $this->db; // For global log_action
+        $this->auth = new Auth($this->db);
     }
 
     public function index() {
-        $this->checkAuth();
-
-        $config = require __DIR__ . '/../../config/database.php';
-        $db = new Database($config);
-        $auth = new Auth($db);
-        $auth->refreshUserSession($_SESSION['user']['id']);
+        $this->checkAuthAndInit();
+        $this->auth->refreshUserSession($_SESSION['user']['id']);
 
         $data = [
             'user' => $_SESSION['user'],
@@ -30,266 +42,137 @@ class SettingsController {
 
         view('settings/index', $data);
     }
-    
-    public function process() {
-        $this->checkAuth();
+
+    // --- NEW: Specific action methods to replace process() ---
+
+    public function updateUsername() {
+        $this->checkAuthAndInit();
         header('Content-Type: application/json');
-        $config = require __DIR__ . '/../../config/database.php';
-        $db = new Database($config);
-        $GLOBALS['db'] = $db;
-        $auth = new Auth($db);
+        
         $userId = $_SESSION['user']['id'];
         $oldUsername = $_SESSION['user']['username'];
+        $newUsername = $_POST['username'];
+
+        if ($oldUsername !== $newUsername) {
+            $this->auth->updateUsername($userId, $newUsername);
+            log_action('INFO', 'SETTINGS_UPDATE', "User updated their username from '{$oldUsername}' to '{$newUsername}'.");
+            echo json_encode(['status' => 'success', 'message' => 'Username updated successfully']);
+        } else {
+            echo json_encode(['status' => 'success', 'message' => 'Username is the same, no changes made.']);
+        }
+        exit;
+    }
+
+    public function updateEmail() {
+        $this->checkAuthAndInit();
+        header('Content-Type: application/json');
+
+        $userId = $_SESSION['user']['id'];
         $oldEmail = $_SESSION['user']['email'] ?? '';
+        $newEmail = trim($_POST['email']);
         
+        if ($oldEmail !== $newEmail) {
+            if ($_SESSION['user']['two_fa'] == 1) {
+                http_response_code(403);
+                echo json_encode(['status' => 'error', 'message' => 'You must disable Two-Factor Authentication before changing your email.']);
+                exit;
+            }
+            
+            $stmt = $this->db->getPdo()->prepare("UPDATE users SET email=? WHERE id=?");
+            $stmt->execute([$newEmail, $userId]);
+            $this->auth->refreshUserSession($userId);
+            log_action('INFO', 'SETTINGS_UPDATE', "User updated their email address from '{$oldEmail}' to '{$newEmail}'.");
+            echo json_encode(['status' => 'success', 'message' => 'Email updated successfully.']);
+        } else {
+            echo json_encode(['status' => 'success', 'message' => 'Email is the same, no changes made.']);
+        }
+        exit;
+    }
+
+    public function updatePassword() {
+        $this->checkAuthAndInit();
+        header('Content-Type: application/json');
+
+        $userId = $_SESSION['user']['id'];
+        $role = $_SESSION['user']['role_name'];
+        $newPassword = $_POST['password'];
+        $submittedOtp = $_POST['otp'] ?? '';
+
         try {
-            $message = 'Settings saved.';
-
-            // --- REMOVED: Unbind Email Logic ---
-            // This is now handled by confirmUnbindEmail()
-            
-            // --- Update Email Logic ---
-            if (isset($_POST['update_email'])) {
-                $newEmail = trim($_POST['email']);
-                if ($oldEmail !== $newEmail) {
-                    // --- SECURITY CHECK: Cannot change email if 2FA is on ---
-                    if ($_SESSION['user']['two_fa'] == 1) {
-                        http_response_code(403);
-                        echo json_encode(['status' => 'error', 'message' => 'You must disable Two-Factor Authentication before changing your email.']);
-                        exit;
-                    }
-                    // --- END CHECK ---
-
-                    $stmt = $db->getPdo()->prepare("UPDATE users SET email=? WHERE id=?");
-                    $stmt->execute([$newEmail, $userId]);
-                    $auth->refreshUserSession($userId);
-                    log_action('INFO', 'SETTINGS_UPDATE', "User updated their email address from '{$oldEmail}' to '{$newEmail}'.");
-                    $message = 'Email updated successfully.';
-                } else {
-                    $message = 'Email is the same, no changes made.';
+            if ($role === 'System Admin') {
+                if (!isset($_SESSION['password_change_otp_required']) || !$_SESSION['password_change_otp_required']) {
+                    throw new Exception('OTP session expired. Please verify current password again.', 403);
                 }
-            }
-            // --- End Update Email Logic ---
-
-            if (isset($_POST['update_username'])) {
-                $newUsername = $_POST['username'];
-                if ($oldUsername !== $newUsername) {
-                    $auth->updateUsername($userId, $newUsername);
-                    log_action('INFO', 'SETTINGS_UPDATE', "User updated their username from '{$oldUsername}' to '{$newUsername}'.");
-                    $message = 'Username updated successfully';
-                } else {
-                    $message = 'Username is the same, no changes made.';
+                if (empty($submittedOtp)) {
+                    throw new Exception('OTP is required to change password.', 400);
                 }
-            }
-            
-            // --- MODIFIED Password Update Logic ---
-            if (isset($_POST['update_password'])) {
-                $role = $_SESSION['user']['role_name'];
-                $newPassword = $_POST['password'];
-                $submittedOtp = $_POST['otp'] ?? '';
-                
-                // 1. Check if OTP is required and validate it for System Admin
-                if ($role === 'System Admin') {
-                    if (!isset($_SESSION['password_change_otp_required']) || !$_SESSION['password_change_otp_required']) {
-                        http_response_code(403);
-                        echo json_encode(['status' => 'error', 'message' => 'OTP session expired. Please verify current password again.']);
-                        exit;
-                    }
-                    if (empty($submittedOtp)) {
-                        http_response_code(400);
-                        echo json_encode(['status' => 'error', 'message' => 'OTP is required to change password.']);
-                        exit;
-                    }
 
-                    // 2. Validate OTP
-                    // NOTE: We reuse Auth::verifyOtp, which also clears the OTP columns upon success.
-                    $otp_result = $auth->verifyOtp($userId, $submittedOtp);
-                    
-                    if (!$otp_result['success']) {
-                        http_response_code(403);
-                        echo json_encode(['status' => 'error', 'message' => $otp_result['message']]);
-                        exit;
-                    }
-                    
-                    // 3. Clear session flags on success
-                    unset($_SESSION['password_change_otp_required']);
-                    unset($_SESSION['password_change_otp_last_sent']);
+                $otp_result = $this->auth->verifyOtp($userId, $submittedOtp);
+                if (!$otp_result['success']) {
+                    throw new Exception($otp_result['message'], 403);
                 }
-                
-                // 4. Update the password
-                $auth->updatePassword($userId, $newPassword);
-                log_action('INFO', 'SETTINGS_UPDATE', "User changed their password." . ($role === 'System Admin' ? " (OTP verified)" : ""));
-                $message = 'Password updated successfully';
+
+                unset($_SESSION['password_change_otp_required']);
+                unset($_SESSION['password_change_otp_last_sent']);
             }
-            // --- END MODIFIED Password Update Logic ---
+
+            $this->auth->updatePassword($userId, $newPassword);
+            log_action('INFO', 'SETTINGS_UPDATE', "User changed their password." . ($role === 'System Admin' ? " (OTP verified)" : ""));
             
-            echo json_encode(['status' => 'success', 'message' => $message]);
+            echo json_encode(['status' => 'success', 'message' => 'Password updated successfully']);
 
         } catch (Exception $e) {
             log_action('ERROR', 'SETTINGS_ERROR', $e->getMessage());
-            http_response_code(500);
-            echo json_encode(['status' => 'error', 'message' => 'An error occurred: ' . $e->getMessage()]);
+            http_response_code($e->getCode() > 0 ? $e->getCode() : 500);
+            echo json_encode(['status' => 'error', 'message' => $e->getMessage()]);
         }
-
         exit;
     }
 
-    /**
-     * NEW: Handles request to unbind email, sends OTP.
-     */
-    public function requestUnbindEmailOtp() {
-        $this->checkAuth();
-        header('Content-Type: application/json');
-        
-        if ($_SESSION['user']['two_fa'] == 1) {
-            http_response_code(403);
-            echo json_encode(['status' => 'error', 'message' => 'You must disable Two-Factor Authentication before removing your email.']);
-            exit;
-        }
-
-        $cooldown_duration = 60; // 60 seconds
-        $last_sent_key = 'unbind_otp_last_sent';
-        $last_sent_time = $_SESSION[$last_sent_key] ?? 0;
-        $time_since_last_sent = time() - $last_sent_time;
-
-        if ($time_since_last_sent < $cooldown_duration) {
-            $remaining_time = $cooldown_duration - $time_since_last_sent;
-            echo json_encode([
-                'status' => 'cooldown',
-                'message' => "Please wait {$remaining_time} seconds before requesting a new code.",
-                'cooldown_remaining' => $remaining_time
-            ]);
-            exit;
-        }
-        
-        $config = require __DIR__ . '/../../config/database.php';
-        $db = new Database($config);
-        $GLOBALS['db'] = $db;
-        $auth = new Auth($db);
-        
-        $userId = $_SESSION['user']['id'];
-        $email = $_SESSION['user']['email'] ?? '';
-
-        if (!empty($email)) {
-            $otp_sent = $auth->generateAndSendOtp($userId, $email);
-            
-            if ($otp_sent) {
-                $_SESSION[$last_sent_key] = time();
-                $_SESSION['unbind_otp_pending'] = true;
-                $message = 'A One-Time Password has been sent to your email to confirm email removal.'; 
-                $status = 'success';
-            } else {
-                $message = 'Failed to send OTP. Check system logs.';
-                $status = 'error';
-            }
-        } else {
-            $message = 'Error: No email to unbind.';
-            $status = 'error';
-        }
-        
-        echo json_encode(['status' => $status, 'message' => $message]);
-        exit;
-    }
-
-    /**
-     * NEW: Verifies OTP and completes email unbinding.
-     */
-    public function confirmUnbindEmail() {
-        $this->checkAuth();
-        header('Content-Type: application/json');
-
-        if (!isset($_SESSION['unbind_otp_pending']) || !$_SESSION['unbind_otp_pending']) {
-            echo json_encode(['status' => 'error', 'message' => 'Invalid session. Please request a new OTP.']);
-            exit;
-        }
-        
-        $config = require __DIR__ . '/../../config/database.php';
-        $db = new Database($config);
-        $GLOBALS['db'] = $db;
-        $auth = new Auth($db);
-
-        $userId = $_SESSION['user']['id'];
-        $submittedOtp = trim($_POST['otp'] ?? '');
-
-        $result = $auth->verifyOtpForToggle($userId, $submittedOtp);
-
-        if ($result['success']) {
-            $currentEmail = $_SESSION['user']['email'] ?? 'none';
-            $stmt = $db->getPdo()->prepare("UPDATE users SET email = NULL WHERE id = ?");
-            $stmt->execute([$userId]);
-            $auth->refreshUserSession($userId);
-            log_action('INFO', 'SETTINGS_UPDATE', "User removed their email address via OTP. (Was: '{$currentEmail}')");
-            
-            unset($_SESSION['unbind_otp_pending']);
-            unset($_SESSION['unbind_otp_last_sent']);
-            
-            echo json_encode(['status' => 'success', 'message' => 'Email address removed successfully.']);
-            exit;
-        } else {
-            echo json_encode(['status' => 'error', 'message' => $result['message']]);
-            exit;
-        }
-    }
-
-
-    /**
-     * MODIFIED: This now acts as STEP 1 for password change.
-     * Checks password, sends OTP for System Admin, or grants direct access for others.
-     */
+    // --- All other methods remain the same ---
+    // (verifyPassword, toggleTwoFA, verifyTwoFAToggleOtp, resendPasswordChangeOtp, 
+    // requestUnbindEmailOtp, confirmUnbindEmail, updateTheme)
+    
     public function verifyPassword() {
-        $this->checkAuth();
+        $this->checkAuthAndInit();
         header('Content-Type: application/json');
         
         $userId = $_SESSION['user']['id'];
         $role = $_SESSION['user']['role_name'];
         $currentPassword = $_POST['current_password'] ?? '';
-        $config = require __DIR__ . '/../../config/database.php';
-        $db = new Database($config);
-        $GLOBALS['db'] = $db;
-        $auth = new Auth($db);
 
-        // 1. Verify the current password regardless of role
-        if (!$auth->verifyPassword($userId, $currentPassword)) {
+        if (!$this->auth->verifyPassword($userId, $currentPassword)) {
             echo json_encode(['status' => 'error', 'message' => 'Incorrect current password.']);
             exit;
         }
 
-        // 2. Check for System Admin and OTP requirement
         if ($role === 'System Admin') {
             $email = $_SESSION['user']['email'] ?? '';
-            
             if (empty($email)) {
                 echo json_encode(['status' => 'error', 'message' => 'System Admin requires an email address set in settings for password changes.']);
                 exit;
             }
             
-            // Send the OTP
-            $otp_sent = $auth->generateAndSendOtp($userId, $email);
+            $otp_sent = $this->auth->generateAndSendOtp($userId, $email);
             
             if (!$otp_sent) {
                 echo json_encode(['status' => 'error', 'message' => 'Failed to send OTP for password change. Check system logs.']);
                 exit;
             }
             
-            // Set session flag to indicate OTP is pending for password change
             $_SESSION['password_change_otp_required'] = true;
-            $_SESSION['password_change_otp_last_sent'] = time(); // Start cooldown timer
+            $_SESSION['password_change_otp_last_sent'] = time();
             
-            // Return status to client to open OTP input modal
             echo json_encode(['status' => 'otp_sent', 'message' => 'Current password verified. OTP sent to email.']);
             exit;
         }
 
-        // 3. For non-Admin users: proceed directly
         echo json_encode(['status' => 'success', 'message' => 'Password verified.']);
         exit;
     }
     
-    /**
-     * NEW: Handles OTP resend for System Admin password change with cooldown.
-     */
     public function resendPasswordChangeOtp() {
-        $this->checkAuth();
+        $this->checkAuthAndInit();
         header('Content-Type: application/json');
 
         if (!isset($_SESSION['password_change_otp_required']) || !$_SESSION['password_change_otp_required']) {
@@ -297,7 +180,7 @@ class SettingsController {
             exit;
         }
 
-        $cooldown_duration = 60; // 60 seconds (1 minute)
+        $cooldown_duration = 60; 
         $last_sent_key = 'password_change_otp_last_sent';
         $last_sent_time = $_SESSION[$last_sent_key] ?? 0;
         $time_since_last_sent = time() - $last_sent_time;
@@ -312,16 +195,11 @@ class SettingsController {
             exit;
         }
         
-        $config = require __DIR__ . '/../../config/database.php';
-        $db = new Database($config);
-        $GLOBALS['db'] = $db;
-        $auth = new Auth($db);
-        
         $userId = $_SESSION['user']['id'];
         $email = $_SESSION['user']['email'] ?? '';
 
         if (!empty($email)) {
-            $otp_sent = $auth->generateAndSendOtp($userId, $email);
+            $otp_sent = $this->auth->generateAndSendOtp($userId, $email);
             
             if ($otp_sent) {
                 $_SESSION[$last_sent_key] = time();
@@ -340,26 +218,14 @@ class SettingsController {
         exit;
     }
     
-    /**
-     * MODIFIED: Handles the request to toggle 2FA on or off.
-     * Initiates OTP flow if DISABLING 2FA.
-     */
     public function toggleTwoFA() {
-        $this->checkAuth();
+        $this->checkAuthAndInit();
         header('Content-Type: application/json');
         
-        $config = require __DIR__ . '/../../config/database.php';
-        $db = new Database($config);
-        $GLOBALS['db'] = $db;
-        $auth = new Auth($db);
-
         $userId = $_SESSION['user']['id'];
         $currentTwoFA = $_SESSION['user']['two_fa'] ?? 0;
-        
-        // Determine the target action
         $targetTwoFA = (int)($_POST['target_two_fa'] ?? 0); 
         
-        // SCENARIO 1: ENABLING 2FA (No OTP needed)
         if ($targetTwoFA == 1 && $currentTwoFA == 0) {
             $email = $_SESSION['user']['email'] ?? '';
             if (empty($email)) {
@@ -367,125 +233,152 @@ class SettingsController {
                  exit;
             }
             
-            $auth->updateTwoFA($userId, 1); 
+            $this->auth->updateTwoFA($userId, 1); 
             log_action('INFO', '2FA_ENABLED', "User #{$userId} successfully enabled 2FA.");
-            echo json_encode([
-                'status' => 'success', 
-                'message' => 'Two-Factor Authentication has been successfully enabled.'
-            ]);
+            echo json_encode(['status' => 'success', 'message' => 'Two-Factor Authentication has been successfully enabled.']);
             exit;
         } 
         
-        // SCENARIO 2: DISABLING 2FA (OTP Required)
         if ($targetTwoFA == 0 && $currentTwoFA == 1) {
             $email = $_SESSION['user']['email'] ?? '';
-
-            // Check Cooldown
             $cooldown_duration = 60; 
             $last_sent_time = $_SESSION['otp_last_sent'] ?? 0;
             $time_since_last_sent = time() - $last_sent_time;
 
-            // Check if user is in a pending state and still in cooldown
             if ($time_since_last_sent < $cooldown_duration && isset($_SESSION['2fa_toggle_pending'])) {
                 $remaining_time = $cooldown_duration - $time_since_last_sent;
-                echo json_encode([
-                    'status' => 'cooldown',
-                    'message' => "Please wait {$remaining_time} seconds before requesting a new code.",
-                    'cooldown_remaining' => $remaining_time
-                ]);
+                echo json_encode(['status' => 'cooldown', 'message' => "Please wait {$remaining_time} seconds.", 'cooldown_remaining' => $remaining_time]);
                 exit;
             }
 
-            // Generate and send OTP
-            $otp_sent = $auth->generateAndSendOtp($userId, $email);
+            $otp_sent = $this->auth->generateAndSendOtp($userId, $email);
 
             if ($otp_sent) {
-                $_SESSION['2fa_toggle_pending'] = true; // Set flag to allow verification later
-                $_SESSION['otp_last_sent'] = time(); // Set cooldown time
+                $_SESSION['2fa_toggle_pending'] = true;
+                $_SESSION['otp_last_sent'] = time();
                 log_action('INFO', '2FA_DISABLE_OTP_SENT', "OTP sent to user #{$userId} to confirm 2FA disablement.");
-                
-                echo json_encode([
-                    'status' => 'otp_required', 
-                    'message' => 'A One-Time Password has been sent to your email to confirm disabling 2FA.'
-                ]);
+                echo json_encode(['status' => 'otp_required', 'message' => 'An OTP has been sent to your email to confirm disabling 2FA.']);
                 exit;
             } else {
-                echo json_encode([
-                    'status' => 'error', 
-                    'message' => 'Failed to send OTP email. Check system configuration.'
-                ]);
+                echo json_encode(['status' => 'error', 'message' => 'Failed to send OTP email.']);
                 exit;
             }
         }
         
-        // SCENARIO 3: NO CHANGE OR INVALID REQUEST
-        $action = $targetTwoFA ? '2FA_ENABLED' : '2FA_DISABLED';
-        // Check if there was an actual change
-        if ($targetTwoFA != $currentTwoFA) {
-            log_action('INFO', $action, "User toggled 2FA status to: " . ($targetTwoFA ? 'Enabled' : 'Disabled'));
-        }
-        echo json_encode(['status' => 'success', 'message' => 'Two-Factor Authentication preference updated.']);
+        echo json_encode(['status' => 'success', 'message' => 'No change in 2FA status.']);
         exit;
     }
     
-    /**
-     * NEW: Verifies the OTP provided to complete the 2FA disabling process.
-     */
     public function verifyTwoFAToggleOtp() {
-        $this->checkAuth();
+        $this->checkAuthAndInit();
         header('Content-Type: application/json');
 
-        if (!isset($_SESSION['user'])) {
-             echo json_encode(['status' => 'error', 'message' => 'Authentication required.']);
-            exit;
-        }
-
         if (!isset($_SESSION['2fa_toggle_pending']) || !$_SESSION['2fa_toggle_pending']) {
-            echo json_encode(['status' => 'error', 'message' => 'Invalid session for OTP verification. Please try toggling 2FA again.']);
+            echo json_encode(['status' => 'error', 'message' => 'Invalid session for OTP verification.']);
             exit;
         }
-
-        $config = require __DIR__ . '/../../config/database.php';
-        $db = new Database($config);
-        $GLOBALS['db'] = $db;
-        $auth = new Auth($db);
 
         $userId = $_SESSION['user']['id'];
         $submittedOtp = trim($_POST['otp'] ?? '');
-
-        $result = $auth->verifyOtpForToggle($userId, $submittedOtp);
+        $result = $this->auth->verifyOtpForToggle($userId, $submittedOtp);
 
         if ($result['success']) {
-            // OTP is correct. Proceed to disable 2FA.
-            $auth->updateTwoFA($userId, 0); // Disable 2FA (set to 0)
+            $this->auth->updateTwoFA($userId, 0); // Disable 2FA
             log_action('INFO', '2FA_DISABLED_OTP', "User #{$userId} successfully disabled 2FA using OTP.");
             
-            unset($_SESSION['2fa_toggle_pending']); // Clear flag
-            unset($_SESSION['otp_last_sent']); // Clear cooldown
+            unset($_SESSION['2fa_toggle_pending']);
+            unset($_SESSION['otp_last_sent']);
             
-            echo json_encode([
-                'status' => 'success', 
-                'message' => 'Two-Factor Authentication has been successfully disabled.'
-            ]);
+            echo json_encode(['status' => 'success', 'message' => 'Two-Factor Authentication has been successfully disabled.']);
             exit;
         } else {
-            echo json_encode([
-                'status' => 'error', 
-                'message' => $result['message']
-            ]);
+            echo json_encode(['status' => 'error', 'message' => $result['message']]);
+            exit;
+        }
+    }
+
+    public function requestUnbindEmailOtp() {
+        $this->checkAuthAndInit();
+        header('Content-Type: application/json');
+        
+        if ($_SESSION['user']['two_fa'] == 1) {
+            http_response_code(403);
+            echo json_encode(['status' => 'error', 'message' => 'You must disable Two-Factor Authentication before removing your email.']);
+            exit;
+        }
+
+        $cooldown_duration = 60;
+        $last_sent_key = 'unbind_otp_last_sent';
+        $last_sent_time = $_SESSION[$last_sent_key] ?? 0;
+        $time_since_last_sent = time() - $last_sent_time;
+
+        if ($time_since_last_sent < $cooldown_duration) {
+            $remaining_time = $cooldown_duration - $time_since_last_sent;
+            echo json_encode(['status' => 'cooldown', 'message' => "Please wait {$remaining_time} seconds.", 'cooldown_remaining' => $remaining_time]);
+            exit;
+        }
+        
+        $userId = $_SESSION['user']['id'];
+        $email = $_SESSION['user']['email'] ?? '';
+
+        if (!empty($email)) {
+            $otp_sent = $this->auth->generateAndSendOtp($userId, $email);
+            
+            if ($otp_sent) {
+                $_SESSION[$last_sent_key] = time();
+                $_SESSION['unbind_otp_pending'] = true;
+                $message = 'An OTP has been sent to your email to confirm removal.'; 
+                $status = 'success';
+            } else {
+                $message = 'Failed to send OTP. Check system logs.';
+                $status = 'error';
+            }
+        } else {
+            $message = 'Error: No email to unbind.';
+            $status = 'error';
+        }
+        
+        echo json_encode(['status' => $status, 'message' => $message]);
+        exit;
+    }
+
+    public function confirmUnbindEmail() {
+        $this->checkAuthAndInit();
+        header('Content-Type: application/json');
+
+        if (!isset($_SESSION['unbind_otp_pending']) || !$_SESSION['unbind_otp_pending']) {
+            echo json_encode(['status' => 'error', 'message' => 'Invalid session. Please request a new OTP.']);
+            exit;
+        }
+        
+        $userId = $_SESSION['user']['id'];
+        $submittedOtp = trim($_POST['otp'] ?? '');
+        $result = $this->auth->verifyOtpForToggle($userId, $submittedOtp);
+
+        if ($result['success']) {
+            $currentEmail = $_SESSION['user']['email'] ?? 'none';
+            $stmt = $this->db->getPdo()->prepare("UPDATE users SET email = NULL WHERE id = ?");
+            $stmt->execute([$userId]);
+            $this->auth->refreshUserSession($userId);
+            log_action('INFO', 'SETTINGS_UPDATE', "User removed their email address via OTP. (Was: '{$currentEmail}')");
+            
+            unset($_SESSION['unbind_otp_pending']);
+            unset($_SESSION['unbind_otp_last_sent']);
+            
+            echo json_encode(['status' => 'success', 'message' => 'Email address removed successfully.']);
+            exit;
+        } else {
+            echo json_encode(['status' => 'error', 'message' => $result['message']]);
             exit;
         }
     }
 
     public function updateTheme() {
-        $this->checkAuth();
+        $this->checkAuthAndInit();
         header('Content-Type: application/json');
         $theme = ($_POST['theme'] ?? 'light') === 'dark' ? 'dark' : 'light';
         $userId = $_SESSION['user']['id'];
-        $config = require __DIR__ . '/../../config/database.php';
-        $db = new Database($config);
-        $auth = new Auth($db);
-        $auth->updateTheme($userId, $theme);
+        $this->auth->updateTheme($userId, $theme);
         echo json_encode(['status' => 'success', 'theme' => $theme]);
         exit;
     }
