+++ app/views/settings/_info.php
@@ -1,0 +1,37 @@
+<div id="tab-info" class="tab-pane">
+    <h3>Web App Information</h3>
+    <div class="info-group">
+        <h4>Application Details</h4>
+        <div class="info-item">
+            <strong>Application Name:</strong>
+            <span>iCensus System</span>
+        </div>
+        <div class="info-item">
+            <strong>Version:</strong>
+            <span>1.0.0</span>
+        </div>
+        <div class="info-item">
+            <strong>Developer:</strong>
+            <span>iCensus Development Team</span>
+        </div>
+    </div>
+    <div class="info-group">
+        <h4>Key Technologies</h4>
+        <div class="info-item">
+            <strong>Backend:</strong>
+            <span>PHP</span>
+        </div>
+            <div class="info-item">
+            <strong>Frontend:</strong>
+            <span>JavaScript, HTML5, CSS3</span>
+        </div>
+        <div class="info-item">
+            <strong>Database:</strong>
+            <span>MySQL</span>
+        </div>
+            <div class="info-item">
+            <strong>Email Service:</strong>
+            <span>PHPMailer v7.0.0</span>
+        </div>
+    </div>
+</div>
